// src/api/openapi/models/TransferInfoRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TransferInfoRequestDto
 */
export interface TransferInfoRequestDto {
    /**
     * 
     * @type {string}
     * @memberof TransferInfoRequestDto
     */
    iban?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TransferInfoRequestDto
     */
    swift?: string | null;
}

/**
 * Check if a given object implements the TransferInfoRequestDto interface.
 */
export function instanceOfTransferInfoRequestDto(value: object): value is TransferInfoRequestDto {
    return true;
}

export function TransferInfoRequestDtoFromJSON(json: any): TransferInfoRequestDto {
    return TransferInfoRequestDtoFromJSONTyped(json, false);
}

export function TransferInfoRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransferInfoRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'iban': json['iban'] == null ? undefined : json['iban'],
        'swift': json['swift'] == null ? undefined : json['swift'],
    };
}

export function TransferInfoRequestDtoToJSON(json: any): TransferInfoRequestDto {
    return TransferInfoRequestDtoToJSONTyped(json, false);
}

export function TransferInfoRequestDtoToJSONTyped(value?: TransferInfoRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'iban': value['iban'],
        'swift': value['swift'],
    };
}